pyproject.toml
[tool.poetry]
name = "kocherga-core"
version = "0.1.0"
description = ""
authors = ["Vyacheslav Matyukhin <slava@kocherga-club.ru>"]

[tool.poetry.dependencies]
python = "^3.7"
slackclient = "*"
markdown = "*"
captcha-solver = "*"
python-dateutil = "*"
"httplib2" = "*"
google-api-python-client = "*"
google-auth = "*"
"google-auth-httplib2" = "*"
gspread = "*"
pillow = "*"
requests = "*"
pyjwt = "*"
mwclient = "*"
termcolor = "*"
pandas = "*"
pyppeteer = { git = "https://github.com/berekuk/pyppeteer.git", branch = "dev" }
attrs = "*"
apscheduler = "*"
fire = "*"
python-gitlab = "*"
slackeventsapi = "*"
pytz = "*"
"boto3" = "*"
mysqlclient = "==1.4.2"
requests-toolbelt = "*"
slappy = "0.4.0"
feedgen = "*"
telethon-sync = "*"
pysocks = "*"
django = "2.2"
django-cors-headers = "*"
django-rest-framework = "*"
sentry-sdk = {extras = ["flask"],version = "*"}
django-debug-toolbar = "*"
react = "*"
honcho = "*"
tqdm = "*"
gunicorn = "^19.9"
aiogram = "^2.0"
aiohttp_socks = "^0.2.2"
django-webpack-loader = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-asyncio = "*"
pytest-env = "*"
pytest-django = "*"
freezegun = "*"
mypy = "*"
black = "==18.6b4"
pycookiecheat = "*"
jupyter = "*"
aioconsole = "*"
matplotlib = "*"
seaborn = "*"
python-docx = "*"
flake8 = "^3.7"
pyinotify = "^0.9.6"
